Backend/requirements.txt
# Zerodha Broker
kiteconnect

# Auth & Login
pyotp
requests

# Config
python-dotenv
pydantic
pydantic-settings

# Web Framework
fastapi
uvicorn[standard]
jinja2

# Data & Indicators
pandas
numpy

# Database
sqlalchemy
psycopg2-binary
asyncpg

# Security
cryptography

# Utilities
apscheduler
